File SDK/pyproject.toml
[tool.poetry]
name = "2025-examples"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flatbuffers = "^24.3.25"
requests = "^2.32.3"
ruamel-yaml = "^0.18.6"
pyzmq = "^26.2.0"
numpy = "<2.0.0"
pygame = "^2.6.1"
opencv-python = "^4.10.0.84"
matplotlib = "^3.9.2"
open3d = "^0.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
